Move get-author, get_label and get_genre to separate module to acheive separation of concerns
